Module capnp::capability

Hooks for for the RPC system.

Roughly corresponds to capability.h in the C++ implementation.

Structs

Client

Params

Promise
This fake Promise struct is defined so that the generated code for interfaces can typecheck even if the "rpc" feature is not enabled.

RemotePromise

Request

Response

Results

Traits

FromClientHook

FromTypelessPipeline

Server