Struct capnp::capability::Client

pub struct Client {
    pub hook: Box<ClientHook>,
}

Fields

hook

Methods

impl Client

fn new(hook: Box<ClientHook>) -> Client

fn new_call<Params, Results>(&self, interface_id: u64, method_id: u16, size_hint: Option<MessageSize>) -> Request<Params, Results>

fn when_resolved(&self) -> Promise<(), Error>

If the capability is actually only a promise, the returned promise resolves once the capability itself has resolved to its final destination (or propagates the exception if the capability promise is rejected). This is mainly useful for error-checking in the case where no calls are being made. There is no reason to wait for this before making calls; if the capability does not resolve, the call results will propagate the error.