Struct capnp::capability::Promise

pub struct Promise<T, E>(_);

This fake Promise struct is defined so that the generated code for interfaces can typecheck even if the "rpc" feature is not enabled.

Methods

impl<T, E> Promise<T, E>

fn ok(v: T) -> Promise<T, E>

fn err(e: E) -> Promise<T, E>

fn map<F, R>(self, _func: F) -> Promise<R, E> where
F: FnOnce(T) -> Result<R, E>,

fn then<F, R>(self, _func: F) -> Promise<R, E> where
F: FnOnce(T) -> Promise<R, E>,