Struct capnp::message::Reader

pub struct Reader<S> where S: ReaderSegments {
    // some fields omitted
}

A container used to read a message.

The underlying implemention uses the ReaderSegments as a trait object. However, we need to include S as concrete type parameter so that the typechecker can correctly deduce appropriate bounds like Send.

Methods

impl<S> Reader<S> where S: ReaderSegments

fn new(segments: S, options: ReaderOptions) -> Reader<S>

fn get_root<'a, T: FromPointerReader<'a>>(&'a self) -> Result<T>

Gets the root of the message, interpreting it as the given type.

fn init_cap_table(&mut self, cap_table: Vec<Option<Box<ClientHook + Send>>>)

fn into_segments(self) -> S