Struct capnp::message::ScratchSpaceHeapAllocator

pub struct ScratchSpaceHeapAllocator<'a> { /* fields omitted */ }

An Allocator whose first segment is a backed by a user-provided buffer.

Recall that an Allocator implementation must ensure that allocated segments are initially zeroed. ScratchSpaceHeapAllocator ensures that is the case by zeroing the entire buffer upon initial construction, and then zeroing any potentially used part of the buffer upon deallocate_segment().

You can reuse a ScratchSpaceHeapAllocator by calling message::Builder::into_allocator(), or by initally passing it to message::Builder::new() as a &mut ScratchSpaceHeapAllocator. Such reuse can save significant amounts of zeroing.

Implementations

impl<'a> ScratchSpaceHeapAllocator<'a>

pub fn new(scratch_space: &'a mut [u8]) -> ScratchSpaceHeapAllocator<'a>

Writes zeroes into the entire buffer and constructs a new allocator from it.

If the buffer is large, this operation could be relatively expensive. If you want to reuse the same scratch space in a later message, you should reuse the entire ScratchSpaceHeapAllocator, to avoid paying this full cost again.

pub fn second_segment_words(self, value: u32) -> ScratchSpaceHeapAllocator<'a>

Sets the size of the second segment in words, where 1 word = 8 bytes. (The first segment is the scratch space passed to ScratchSpaceHeapAllocator::new().

pub fn allocation_strategy(
self,
value: AllocationStrategy
) -> ScratchSpaceHeapAllocator<'a>

Sets the allocation strategy for segments after the second one.