Trait capnp::message::MessageBuilder

pub trait MessageBuilder {
    fn arena_mut(&mut self) -> &mut BuilderArena;
    fn arena(&self) -> &BuilderArena;

    fn get_root_internal<'a>(&mut self) -> Builder<'a> { ... }
    fn init_root<'a, T: FromPointerBuilder<'a>>(&'a mut self) -> T { ... }
    fn get_root<'a, T: FromPointerBuilder<'a>>(&'a mut self) -> Result<T> { ... }
    fn set_root<To, From: SetPointerBuilder<To>>(&mut self, value: From) -> Result<()> { ... }
    fn get_segments_for_output<'a>(&'a self) -> OutputSegments<'a> { ... }
    fn get_cap_table<'a>(&'a self) -> &'a [Option<Box<ClientHook + Send>>] { ... }
}

An abstract container used to build a message.

Required Methods

fn arena_mut(&mut self) -> &mut BuilderArena

fn arena(&self) -> &BuilderArena

Provided Methods

fn get_root_internal<'a>(&mut self) -> Builder<'a>

fn init_root<'a, T: FromPointerBuilder<'a>>(&'a mut self) -> T

Initializes the root as a value of the given type.

fn get_root<'a, T: FromPointerBuilder<'a>>(&'a mut self) -> Result<T>

Gets the root, interpreting it as the given type.

fn set_root<To, From: SetPointerBuilder<To>>(&mut self, value: From) -> Result<()>

Sets the root to a deep copy of the given value.

fn get_segments_for_output<'a>(&'a self) -> OutputSegments<'a>

fn get_cap_table<'a>(&'a self) -> &'a [Option<Box<ClientHook + Send>>]