Struct capnp::private::AlignedData

pub struct AlignedData<T> {
    pub _dummy: u64,
    pub data: T,
}

Some data that's guaranteed to be aligned on a word boundary.

Typically the type parameter T will be instantiated as [u8; n] where n is a multiple of eight.

Perhaps in the future Rust will provide a nicer way to guarantee alignment.

DEPRECATED. Should use capnp_word!() instead.

Fields

_dummy

data