Module capnp::raw

Functions providing low level access to encoded data.

Functions

get_list_bytes
Gets a slice view of a list, exluding any tag word.

get_list_element_size
Gets the size of the elements in a list.

get_list_step_size_in_bits
Gets the number of bits between successive elements in a list.

get_struct_data_section
Gets a slice view of the data section of a struct.

get_struct_pointer_section
Gets the pointer section as a list.