Enum capnp::serialize::ReadContinuation

pub enum ReadContinuation {
    SegmentTableFirst {
        buf: [u8; 8],
        idx: usize,
    },
    SegmentTableRest {
        segment_count: usize,
        segment_slices: Vec<(usize, usize)>,
        buf: [u8; 8],
        idx: usize,
    },
    Segments {
        segment_slices: Vec<(usize, usize)>,
        owned_space: Vec<Word>,
        idx: usize,
    },
}

Variants

SegmentTableFirst

Reading the message would block while trying to read the first word (the segment count, and the first segment's length).

Fields

buf

The buffer being read into.

idx

The number of bytes read before being blocked.

SegmentTableRest

Reading the message would block while trying to read the rest of the segment table.

Fields

segment_count

The total number of segments.

segment_slices

The segment start and end offsets into the segment buffer.

buf

The buffer being read into.

idx

The number of bytes read before being blocked.

Segments

Reading the message would block while trying to read the segments.

Fields

segment_slices

The segment start and end offsets into the segment buffer.

owned_space

The segment buffer.

idx

The number of bytes read into owned_space before being blocked.