Enum capnp::serialize::WriteContinuation

pub enum WriteContinuation {
    SegmentTable {
        word: usize,
        idx: usize,
    },
    Segments {
        segment: usize,
        idx: usize,
    },
}

Variants

SegmentTable

Writing the message would block while trying to write the segment table.

Fields

word

The next word of the segment table to write.

idx

The byte offset into the next word to write.

Segments

Writing the message would block while trying to write the message segments.

Fields

segment

The next segment to write.

idx

The byte offset into the next segment to write.