Function capnp::serialize::read_message_async

pub fn read_message_async<R>(read: &mut R, options: ReaderOptions, continuation: Option<ReadContinuation>) -> Result<AsyncRead> where R: Read

Reads a Cap'n Proto serialized message from read with the provided options.

Takes an optional continuation value from a prior blocked message read attempt.

For optimal performance, read should be a buffered reader type.

Returns a Complete(OwnedSpaceMessageReader) if the message read succeeds, or a Continue(ReadContinuation) if the read would block. The caller should use the continuation to resume reading the message again when read can supply more bytes.