Struct capnp::Word

#[repr(C)]

pub struct Word { /* fields omitted */ }

Eight bytes of memory with opaque interior. Use capnp_word!() to construct one of these.

This type is used to ensure that the data of a message is properly aligned.

Methods

impl Word

pub fn allocate_zeroed_vec(length: usize) -> Vec<Word>

Does this, but faster: ::std::iter::repeat(Word(0)).take(length).collect()

pub unsafe fn bytes_to_words<'a>(bytes: &'a [u8]) -> &'a [Word]

Converts a byte slice into a Word slice. Unsafe due to possible alignment issues. Only call this if you know that either

bytes.to_ptr() falls on an eight-byte boundary, or

your processor is okay with unaligned reads.

pub unsafe fn bytes_to_words_mut<'a>(bytes: &'a mut [u8]) -> &'a mut [Word]

Converts a mutable byte slice into a mutable Word slice. Unsafe due to possible alignment issues. Only call this if you know that either

bytes.to_ptr() falls on an eight-byte boundary, or

your processor is okay with unaligned reads and writes

pub fn words_to_bytes<'a>(words: &'a [Word]) -> &'a [u8]

pub fn words_to_bytes_mut<'a>(words: &'a mut [Word]) -> &'a mut [u8]