Module capnp::capability

Hooks for for the RPC system.

Roughly corresponds to capability.h in the C++ implementation.

Structs

Client
An untyped client.

Params
The values of the parameters passed to a method call, as seen by the server.

Promise
A computation that might eventually resolve to a value of type T or to an error of type E. Dropping the promise cancels the computation.

RemotePromise
A promise for a result from a method call.

Request
A method call that has not been sent yet.

Response
A response from a method call, as seen by the client.

Results
The return values of a method, written in-place by the method body.

Traits

FromClientHook

FromServer
Trait to track the relationship between generated Server traits and Client structs.

FromTypelessPipeline

Server
An untyped server.