Module capnp::message

Untyped root container for a Cap’n Proto value.

Structs

Builder
A container used to build a message.

HeapAllocator
Standard segment allocator. Allocates each segment via alloc::alloc::alloc_zeroed().

Reader
A container used to read a message.

ReaderOptions
Options controlling how data is read.

ScratchSpaceHeapAllocator
An Allocator whose first segment is a backed by a user-provided buffer.

SegmentArray
An array of segments.

TypedReader
A message reader whose value is known to be of type T.

Enums

AllocationStrategy

Constants

DEFAULT_READER_OPTIONS

SUGGESTED_ALLOCATION_STRATEGY

SUGGESTED_FIRST_SEGMENT_WORDS

Traits

Allocator
An object that allocates memory for a Cap’n Proto message as it is being built.

ReaderSegments
An object that manages the buffers underlying a Cap’n Proto message reader.