Module capnp::serialize

Reading and writing of messages using the standard stream framing.

Structs

OwnedSegments

SliceSegments
Segments read from a single flat slice of words.

Functions

compute_serialized_size_in_words
Returns the number of words required to serialize the message.

read_message
Reads a serialized message from a stream with the provided options.

read_message_from_words
Reads a serialized message from a slice of words.

write_message
Writes the provided message to write.

write_message_segments

write_message_segments_to_words

write_message_to_words
Constructs a flat vector containing the entire message.