Module capnp::traits

Structs

ListIter

Traits

FromPointerBuilder

FromPointerBuilderRefDefault

FromPointerReader

FromPointerReaderRefDefault

FromStructBuilder

FromStructReader

FromU16

HasStructSize

HasTypeId

Imbue

ImbueMut

IndexMove

Owned
Associated types hackery that allows us to reason about Cap'n Proto types without needing to give them a lifetime 'a.

OwnedStruct

Pipelined

SetPointerBuilder

ToU16