Module capnp::traits

Structs

ListIter

Traits

FromPointerBuilder

FromPointerReader

FromStructBuilder

FromStructReader

FromU16

HasStructSize

HasTypeId

Imbue

ImbueMut

IndexMove

IntoInternalListReader

IntoInternalStructReader

Owned
Associated types hackery that allows us to reason about Cap’n Proto types without needing to give them a lifetime 'a.

OwnedStruct

Pipelined

SetPointerBuilder

ToU16